+++ package.json
@@ -28,5 +28,5 @@
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.6.1"
   },
-  "proxy": "http://localhost:3004"
+  "proxy": "http://localhost:8080"
 }
