+++ src/Components/HomePage/Homepage.css
@@ -1,0 +1,24 @@
+.HomePage-Container{
+  display: flex;
+  flex-grow: 1;
+  border: 2px solid black;
+}
+
+.HomePage-signup{
+  display: flex;
+  flex-grow: 1;
+  border: 1px solid black;
+}
+
+.HomePage-login{
+  display: flex;
+  flex-grow: 1;
+  border: 1px solid black;
+}
+
+.HomePage-divide{
+  display: flex;
+  width: 5px;
+  border: 1px solid black;
+  background: rgb(0, 0, 0);
+}
