+++ src/Components/HomePage/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './Homepage.css';
+import Signup from '../Signup';
+import Login from '../Login';
+
+const HomePage = props => (
+  <div className="HomePage-Container">
+    <div className="HomePage-signup"><Signup /></div>
+    <div className="HomePage-divide">b</div>
+    <div className="HomePage-login"><Login /></div>
+  </div>
+);
+
+export default HomePage;
