+++ src/Components/Render/index.js
@@ -2,14 +2,14 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Header from '../Header';
 import Footer from '../Footer';
-import Signup from '../Signup';
+import HomePage from '../HomePage';
 
 class Main extends React.Component {
   render() {
     return (
       <div>
         <Header />
-        <Signup />
+        <HomePage />
         <Footer />
       </div>
     );
