+++ src/Components/Signup/Signup.css
@@ -1,7 +1,27 @@
-.Signup-row{
-  padding: 5px;
+.Signup-box{
+  margin: auto;
+  border: 1px solid black;
+  border-radius: 3%;
+  padding: 3%;
+  display: flex;
+  flex-direction: column;
 }
 
-.paddingTop{
-  margin-top: 20px;
+.Signup-redirect-button{
+  background-color: rgb(45, 45, 45);
+  margin: auto;
+  padding: 20px;
+  outline: none;
+}
+
+.Signup-button-label{
+  color:rgb(240, 240, 240);
+  font-size: 20px;
+  margin: auto;
+}
+
+.Signup-catchyline{
+  text-align: center;
+  font-family: 'Helvetica', sans-serif;
+  font-size: 25px;
 }
